# docs/frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  version: 1.0.0
  description: 'Personal expenses, bills split between people, balances and settling
    up, CSV exports and bill scanning. Authenticate with `POST /api/v1/auth/login/`
    and send `Authorization: Bearer <access>`.'
paths:
...
  /api/v1/expenses/:
    get:
      operationId: expenses_list
      description: All time unless `start`/`end` narrow it. `count` and `total_amount`
        cover the whole filtered set, not just this page. Follow `next` for more.
      summary: List expenses, newest date first
      parameters:
      - in: query
        name: category
        schema:
          type: integer
        description: A category id.
      - name: cursor
        required: false
        in: query
        description: The pagination cursor value.
        schema:
          type: string
      - in: query
        name: end
        schema:
          type: string
          format: date
        description: On or before (YYYY-MM-DD).
      - in: query
        name: page_size
        schema:
          type: integer
        description: 1-100, default 25.
      - in: query
        name: search
        schema:
          type: string
        description: Matches the note, line items and people.
      - in: query
        name: start
        schema:
          type: string
          format: date
        description: On or after (YYYY-MM-DD).
      tags:
      - expenses
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExpensePageList'
          description: ''
    post:
      operationId: expenses_create
      description: |-
        Scope every queryset to the requester, and stamp ownership on create.

        The direct analogue of OwnerScopedMixin and OwnerFormMixin. `user` is
        never a writable field, so ownership cannot be reassigned by a payload.
      summary: Create an expense
      tags:
      - expenses
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ExpenseRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'
          description: ''
...
components:
  schemas:
...
    Expense:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        category:
          type: integer
        category_name:
          type: string
          readOnly: true
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        spent_on:
          type: string
          format: date
        note:
          type: string
          maxLength: 255
        paid_by:
          type: integer
          nullable: true
        paid_by_name:
          type: string
          readOnly: true
          nullable: true
        participants:
          type: array
          items:
            type: integer
        misc_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        misc_note:
          type: string
          maxLength: 100
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseItem'
        items_total:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
        unaccounted_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
          nullable: true
          readOnly: true
        is_balanced:
          type: boolean
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount
      - category
      - category_name
      - created_at
      - id
      - is_balanced
      - items_total
      - paid_by_name
      - spent_on
      - unaccounted_amount
    ExpenseItem:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        shares:
          type: array
          items:
            $ref: '#/components/schemas/ItemShare'
      required:
      - amount
      - id
      - name
    ExpenseItemRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        shares:
          type: array
          items:
            $ref: '#/components/schemas/ItemShareRequest'
      required:
      - amount
      - name
    ExpensePage:
      type: object
      properties:
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        count:
          type: integer
        total_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        results:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
      required:
      - count
      - next
      - previous
      - results
      - total_amount
...
    ExpenseRequest:
      type: object
      properties:
        category:
          type: integer
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
        spent_on:
          type: string
          format: date
        note:
          type: string
          maxLength: 255
        paid_by:
          type: integer
          nullable: true
        participants:
          type: array
          items:
            type: integer
        misc_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        misc_note:
          type: string
          maxLength: 100
        include_self:
          type: boolean
          writeOnly: true
          default: true
        bill_scan:
          type: integer
          writeOnly: true
          nullable: true
          description: 'On create only: the scan this expense was confirmed from.'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseItemRequest'
      required:
      - amount
      - category
      - spent_on
...
    ItemShare:
      type: object
      properties:
        participant:
          type: integer
        weight:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
      required:
      - participant
    ItemShareRequest:
      type: object
      properties:
        participant:
          type: integer
        weight:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
      required:
      - participant
...
    PaginatedExpensePageList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/ExpensePage'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT